fix module export
